Conflicts:
	setup.sh

+++ setup.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
-eeros_branch="0.3.x"
-eeduro_branch="develop"
+. lib.sh.in
+
+setc eeros_branch "master"
+setc eeduro_branch "master"
+setc toolchain ""
 
 missing_packages=""
 missing_package_count=0
@@ -45,20 +48,32 @@ fi
 
 if [ ! -d eeros ]; then
 	git clone git://github.com/eeros-project/eeros-framework.git eeros -o upstream -b $eeros_branch
+	if [ $? -ne 0 ]; then
+		exit 2
+	fi
 fi
 
 if [ ! -d eeduro ]; then
 	git clone git://github.com/eeduro/eeduro-platform.git eeduro -o upstream -b $eeduro_branch
+	if [ $? -ne 0 ]; then
+		exit 3
+	fi
 fi
 
-if [ ! -d linaro-tc ]; then
-	wget -c https://launchpad.net/linaro-toolchain-binaries/trunk/2013.03/+download/gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux.tar.bz2
-	echo extracting...
-	tar xjf gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux.tar.bz2
-	rm gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux.tar.bz2
-	mv gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux linaro-tc
-	cp toolchain.cmake linaro-tc/
+if [ -z $toolchain ]; then
+	toolchain="../linaro-tc/toolchain.cmake"
+	if [ ! -d linaro-tc ]; then
+		wget -c https://launchpad.net/linaro-toolchain-binaries/trunk/2013.03/+download/gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux.tar.bz2
+		if [ $? -ne 0 ]; then
+			exit 4
+		fi
+		echo extracting...
+		tar xjf gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux.tar.bz2
+		rm gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux.tar.bz2
+		mv gcc-linaro-arm-linux-gnueabihf-4.7-2013.03-20130313_linux linaro-tc
+		cp toolchain.cmake linaro-tc/
+	fi
 fi
 
 create_build_dir build-x86-64
-create_build_dir build-armhf -DCMAKE_TOOLCHAIN_FILE=../linaro-tc/toolchain.cmake
+create_build_dir build-armhf -DCMAKE_TOOLCHAIN_FILE=$toolchain
